:construction: Rework variable handling
The entries in apt sources files are objects of a defined type which can
completely be represented in yaml aka ansible variables.  The templates
can be simplified a lot then, removing all kinds of convoluted and
complicated logic by simple loops over those objects.  Those objects are
assembled in 'vars' now while defaults are more or less the same as
before, just extending a little bit while keeping it overall clean and
understandable.  Support for setups with proxies and/or standard URIs is
improved, especially for mixed setups.